clean up test
